Add test for missing parameters
