+++ GraphLinqQL.Resolvers/WellKnownErrorCodes.cs
@@ -17,6 +17,7 @@ namespace GraphLinqQL
         public const string UndefinedField = "undefinedField";
         public const string ParseError = "parseError";
         public const string ErrorInResolver = "errorInResolver";
+        public const string MissingArgument = "missingArgument";
 
         private static readonly IReadOnlyDictionary<string, MessageResolver> messages = ConstructMessages();
 
@@ -29,6 +30,7 @@ namespace GraphLinqQL
             builder.Add(UndefinedField, args => string.Format(CultureInfo.InvariantCulture, ErrorMessages.ErrorMessages.undefinedField, args["fieldName"], args["type"]));
             builder.Add(ParseError, args => string.Format(CultureInfo.InvariantCulture, ErrorMessages.ErrorMessages.parseError, args["innerMessage"]));
             builder.Add(ErrorInResolver, args => string.Format(CultureInfo.InvariantCulture, ErrorMessages.ErrorMessages.errorInResolver, args["fieldName"], args["type"]));
+            builder.Add(MissingArgument, args => string.Format(CultureInfo.InvariantCulture, ErrorMessages.ErrorMessages.missingArgument, args["fieldName"], args["argument"]));
             return builder.ToImmutable();
         }
